Top and Bottom Views of a tree
The idea is to do something similar to vertical Order Traversal. Like
vertical Order Traversal, we need to nodes of same horizontal distance
together. We do a level order traversal so that the topmost node at a
horizontal node is visited before any other node of same horizontal
distance below it. Hashing is used to check if a node at given
horizontal distance is seen or not.